requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
boto3==1.34.15
botocore==1.34.15
certifi==2023.11.17
cffi==1.16.0
Django==4.2.7
django-cors-headers==4.3.0
django-minio-storage==0.5.7
django-redis==5.4.0
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
minio==7.2.3
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
